+++ zellij-client/src/lib.rs
@@ -3,6 +3,7 @@ pub mod os_input_output;
 mod command_is_executing;
 mod input_handler;
 
+use log::info;
 use std::env::current_exe;
 use std::io::{self, Write};
 use std::path::Path;
@@ -88,6 +89,7 @@ pub fn start_client(
     info: ClientInfo,
     layout: Option<Layout>,
 ) {
+    info!("Starting Zellij client!");
     let clear_client_terminal_attributes = "\u{1b}[?1l\u{1b}=\u{1b}[r\u{1b}12l\u{1b}[?1000l\u{1b}[?1002l\u{1b}[?1003l\u{1b}[?1005l\u{1b}[?1006l\u{1b}[?12l";
     let take_snapshot = "\u{1b}[?1049h";
     let bracketed_paste = "\u{1b}[?2004h";
@@ -118,7 +120,6 @@ pub fn start_client(
         palette,
     };
 
-    #[cfg(not(any(feature = "test", test)))]
     let first_msg = match info {
         ClientInfo::Attach(name, force, config_options) => {
             SESSION_NAME.set(name).unwrap();
@@ -140,16 +141,6 @@ pub fn start_client(
             )
         }
     };
-    #[cfg(any(feature = "test", test))]
-    let first_msg = {
-        let _ = SESSION_NAME.set("".into());
-        ClientToServerMsg::NewClient(
-            client_attributes,
-            Box::new(opts),
-            Box::new(config_options.clone()),
-            layout,
-        )
-    };
 
     os_input.connect_to_server(&*ZELLIJ_IPC_PIPE);
     os_input.send_to_server(first_msg);
@@ -167,7 +158,6 @@ pub fn start_client(
     > = channels::bounded(50);
     let send_client_instructions = SenderWithContext::new(send_client_instructions);
 
-    #[cfg(not(any(feature = "test", test)))]
     std::panic::set_hook({
         use zellij_utils::errors::handle_panic;
         let send_client_instructions = send_client_instructions.clone();
@@ -175,6 +165,8 @@ pub fn start_client(
             handle_panic(info, &send_client_instructions);
         })
     });
+
+    let on_force_close = config_options.on_force_close.unwrap_or_default();
 
     let _stdin_thread = thread::Builder::new()
         .name("stdin_handler".to_string())
@@ -212,7 +204,7 @@ pub fn start_client(
                     Box::new({
                         let os_api = os_input.clone();
                         move || {
-                            os_api.send_to_server(ClientToServerMsg::Action(Action::Detach));
+                            os_api.send_to_server(ClientToServerMsg::Action(on_force_close.into()));
                         }
                     }),
                 );
@@ -303,6 +295,7 @@ pub fn start_client(
     );
 
     os_input.disable_mouse();
+    info!("{}", exit_msg);
     os_input.unset_raw_mode(0);
     let mut stdout = os_input.get_stdout_writer();
     let _ = stdout.write(goodbye_message.as_bytes()).unwrap();