+++ zellij-server/src/screen.rs
@@ -15,10 +15,10 @@ use crate::{
     wasm_vm::PluginInstruction,
     ServerInstruction, SessionState,
 };
-use zellij_tile::data::{Event, InputMode, ModeInfo, Palette, PluginCapabilities, TabInfo};
+use zellij_tile::data::{Event, ModeInfo, Palette, PluginCapabilities, TabInfo};
 use zellij_utils::{
     errors::{ContextType, ScreenContext},
-    input::options::Options,
+    input::{get_mode_info, options::Options},
     ipc::ClientAttributes,
     pane_size::PositionAndSize,
 };
@@ -50,16 +50,17 @@ pub(crate) enum ScreenInstruction {
     ScrollUpAt(Position),
     ScrollDown,
     ScrollDownAt(Position),
+    ScrollToBottom,
     PageScrollUp,
     PageScrollDown,
     ClearScroll,
     CloseFocusedPane,
     ToggleActiveTerminalFullscreen,
     TogglePaneFrames,
-    SetSelectable(PaneId, bool),
-    SetFixedHeight(PaneId, usize),
-    SetFixedWidth(PaneId, usize),
-    SetInvisibleBorders(PaneId, bool),
+    SetSelectable(PaneId, bool, usize),
+    SetFixedHeight(PaneId, usize, usize),
+    SetFixedWidth(PaneId, usize, usize),
+    SetInvisibleBorders(PaneId, bool, usize),
     ClosePane(PaneId),
     ApplyLayout(Layout, Vec<RawFd>),
     NewTab(RawFd),
@@ -104,6 +105,7 @@ impl From<&ScreenInstruction> for ScreenContext {
             ScreenInstruction::Exit => ScreenContext::Exit,
             ScreenInstruction::ScrollUp => ScreenContext::ScrollUp,
             ScreenInstruction::ScrollDown => ScreenContext::ScrollDown,
+            ScreenInstruction::ScrollToBottom => ScreenContext::ScrollToBottom,
             ScreenInstruction::PageScrollUp => ScreenContext::PageScrollUp,
             ScreenInstruction::PageScrollDown => ScreenContext::PageScrollDown,
             ScreenInstruction::ClearScroll => ScreenContext::ClearScroll,
@@ -151,7 +153,6 @@ pub(crate) struct Screen {
     /// The index of this [`Screen`]'s active [`Tab`].
     active_tab_index: Option<usize>,
     mode_info: ModeInfo,
-    input_mode: InputMode,
     colors: Palette,
     session_state: Arc<RwLock<SessionState>>,
     draw_pane_frames: bool,
@@ -164,7 +165,6 @@ impl Screen {
         client_attributes: &ClientAttributes,
         max_panes: Option<usize>,
         mode_info: ModeInfo,
-        input_mode: InputMode,
         session_state: Arc<RwLock<SessionState>>,
         draw_pane_frames: bool,
     ) -> Self {
@@ -176,7 +176,6 @@ impl Screen {
             active_tab_index: None,
             tabs: BTreeMap::new(),
             mode_info,
-            input_mode,
             session_state,
             draw_pane_frames,
         }
@@ -197,7 +196,6 @@ impl Screen {
             self.max_panes,
             Some(PaneId::Terminal(pane_id)),
             self.mode_info.clone(),
-            self.input_mode,
             self.colors,
             self.session_state.clone(),
             self.draw_pane_frames,
@@ -345,6 +343,11 @@ impl Screen {
         }
     }
 
+    /// Returns a mutable reference to this [`Screen`]'s indexed [`Tab`].
+    pub fn get_indexed_tab_mut(&mut self, tab_index: usize) -> Option<&mut Tab> {
+        self.get_tabs_mut().get_mut(&tab_index)
+    }
+
     /// Creates a new [`Tab`] in this [`Screen`], applying the specified [`Layout`]
     /// and switching to it.
     pub fn apply_layout(&mut self, layout: Layout, new_pids: Vec<RawFd>) {
@@ -360,12 +363,11 @@ impl Screen {
             self.max_panes,
             None,
             self.mode_info.clone(),
-            self.input_mode,
             self.colors,
             self.session_state.clone(),
             self.draw_pane_frames,
         );
-        tab.apply_layout(layout, new_pids);
+        tab.apply_layout(layout, new_pids, tab_index);
         self.active_tab_index = Some(tab_index);
         self.tabs.insert(tab_index, tab);
         self.update_tabs();
@@ -437,21 +439,18 @@ pub(crate) fn screen_thread_main(
 ) {
     let capabilities = config_options.simplified_ui;
     let draw_pane_frames = !config_options.no_pane_frames;
-    let default_mode = config_options.default_mode.unwrap_or_default();
 
     let mut screen = Screen::new(
         bus,
         &client_attributes,
         max_panes,
-        ModeInfo {
-            palette: client_attributes.palette,
-            capabilities: PluginCapabilities {
+        get_mode_info(
+            config_options.default_mode.unwrap_or_default(),
+            client_attributes.palette,
+            PluginCapabilities {
                 arrow_fonts: capabilities,
             },
-            mode: default_mode,
-            ..ModeInfo::default()
-        },
-        default_mode,
+        ),
         session_state,
         draw_pane_frames,
     );
@@ -587,6 +586,12 @@ pub(crate) fn screen_thread_main(
                     .unwrap()
                     .scroll_terminal_down(&point, 3);
             }
+            ScreenInstruction::ScrollToBottom => {
+                screen
+                    .get_active_tab_mut()
+                    .unwrap()
+                    .scroll_active_terminal_to_bottom();
+            }
             ScreenInstruction::PageScrollUp => {
                 screen
                     .get_active_tab_mut()
@@ -609,29 +614,53 @@ pub(crate) fn screen_thread_main(
                 screen.get_active_tab_mut().unwrap().close_focused_pane();
                 screen.render();
             }
-            ScreenInstruction::SetSelectable(id, selectable) => {
-                screen
-                    .get_active_tab_mut()
-                    .unwrap()
-                    .set_pane_selectable(id, selectable);
+            ScreenInstruction::SetSelectable(id, selectable, tab_index) => {
+                screen.get_indexed_tab_mut(tab_index).map_or_else(
+                    || {
+                        log::warn!(
+                            "Tab index #{} not found, could not set selectable for plugin #{:?}.",
+                            tab_index,
+                            id
+                        )
+                    },
+                    |tab| tab.set_pane_selectable(id, selectable),
+                );
             }
-            ScreenInstruction::SetFixedHeight(id, fixed_height) => {
-                screen
-                    .get_active_tab_mut()
-                    .unwrap()
-                    .set_pane_fixed_height(id, fixed_height);
+            ScreenInstruction::SetFixedHeight(id, fixed_height, tab_index) => {
+                screen.get_indexed_tab_mut(tab_index).map_or_else(
+                    || {
+                        log::warn!(
+                            "Tab index #{} not found, could not set fixed height for plugin #{:?}.",
+                            tab_index,
+                            id
+                        )
+                    },
+                    |tab| tab.set_pane_fixed_height(id, fixed_height),
+                );
             }
-            ScreenInstruction::SetFixedWidth(id, fixed_width) => {
-                screen
-                    .get_active_tab_mut()
-                    .unwrap()
-                    .set_pane_fixed_width(id, fixed_width);
+            ScreenInstruction::SetFixedWidth(id, fixed_width, tab_index) => {
+                screen.get_indexed_tab_mut(tab_index).map_or_else(
+                    || {
+                        log::warn!(
+                            "Tab index #{} not found, could not set fixed width for plugin #{:?}.",
+                            tab_index,
+                            id
+                        )
+                    },
+                    |tab| tab.set_pane_fixed_width(id, fixed_width),
+                );
             }
-            ScreenInstruction::SetInvisibleBorders(id, invisible_borders) => {
-                screen
-                    .get_active_tab_mut()
-                    .unwrap()
-                    .set_pane_invisible_borders(id, invisible_borders);
+            ScreenInstruction::SetInvisibleBorders(id, invisible_borders, tab_index) => {
+                screen.get_indexed_tab_mut(tab_index).map_or_else(
+                    || {
+                        log::warn!(
+                            r#"Tab index #{} not found, could not set invisible borders for plugin #{:?}."#,
+                            tab_index,
+                            id
+                        )
+                    },
+                    |tab| tab.set_pane_invisible_borders(id, invisible_borders),
+                );
                 screen.render();
             }
             ScreenInstruction::ClosePane(id) => {