+++ zellij-utils/src/input/actions.rs
@@ -1,6 +1,7 @@
 //! Definition of the actions that can be bound to keys.
 
 use super::command::RunCommandAction;
+use crate::input::options::OnForceClose;
 use serde::{Deserialize, Serialize};
 use zellij_tile::data::InputMode;
 
@@ -47,6 +48,8 @@ pub enum Action {
     ScrollDown,
     /// Scroll down at point
     ScrollDownAt(Position),
+    /// Scroll down to bottom in focus pane.
+    ScrollToBottom,
     /// Scroll up one page in focus pane.
     PageScrollUp,
     /// Scroll down one page in focus pane.
@@ -82,4 +85,13 @@ pub enum Action {
     MouseRelease(Position),
     MouseHold(Position),
     Copy,
+}
+
+impl From<OnForceClose> for Action {
+    fn from(ofc: OnForceClose) -> Action {
+        match ofc {
+            OnForceClose::Quit => Action::Quit,
+            OnForceClose::Detach => Action::Detach,
+        }
+    }
 }